Fix build target to make importing easier (#5)
The change in build target was made to allow for BitInt literals, but altered behavior in a way that makes importing a much larger headache. This commit reverts the change and explicitly constructs BigInts in all tests.

It also adds yarn to DevDependencies as it otherwise required developers to install it globally to run any tests.